requirements.txt
kaggle
google-cloud-storage